+++ pglast/printers/dml.py
@@ -3,7 +3,7 @@
 # :Created:   sab 05 ago 2017 16:34:08 CEST
 # :Author:    Lele Gaifax <lele@metapensiero.it>
 # :License:   GNU General Public License version 3 or later
-# :Copyright: © 2017, 2018, 2019 Lele Gaifax
+# :Copyright: © 2017, 2018, 2019, 2020 Lele Gaifax
 #
 
 from .. import enums
@@ -731,6 +731,17 @@ def row_expr(node, output):
                                   coerciontype)
 
 
+def _select_needs_to_be_wrapped_in_parens(node):
+    # Accordingly with https://www.postgresql.org/docs/current/sql-select.html, a SELECT
+    # statement on either sides of UNION/INTERSECT/EXCEPT must be wrapped in parens if it
+    # contains ORDER BY/LIMIT/... or is a nested UNION/INTERSECT/EXCEPT
+    return (node.sortClause
+            or node.limitCount
+            or node.limitOffset
+            or node.lockingClause
+            or node.op != enums.SetOperation.SETOP_NONE)
+
+
 @node_printer('SelectStmt')
 def select_stmt(node, output):
     with output.push_indent():
@@ -741,6 +752,8 @@ def select_stmt(node, output):
             output.space(2)
             output.indent()
 
+        so = enums.SetOperation
+
         if node.valuesLists:
             # Is this a SELECT ... FROM (VALUES (...))?
             require_parens = node.parent_node.node_tag == 'RangeSubselect'
@@ -750,12 +763,17 @@ def select_stmt(node, output):
             output.print_lists(node.valuesLists)
             if require_parens:
                 output.write(')')
-        elif node.targetList is Missing:
+        elif node.op != so.SETOP_NONE and (node.larg or node.rarg):
             with output.push_indent():
-                output.print_node(node.larg)
+                if node.larg:
+                    if _select_needs_to_be_wrapped_in_parens(node.larg):
+                        output.write('(')
+                        output.print_node(node.larg)
+                        output.write(')')
+                    else:
+                        output.print_node(node.larg)
                 output.newline()
                 output.newline()
-                so = enums.SetOperation
                 if node.op == so.SETOP_UNION:
                     output.write('UNION')
                 elif node.op == so.SETOP_INTERSECT:
@@ -766,7 +784,13 @@ def select_stmt(node, output):
                     output.write(' ALL')
                 output.newline()
                 output.newline()
-                output.print_node(node.rarg)
+                if node.rarg:
+                    if _select_needs_to_be_wrapped_in_parens(node.rarg):
+                        output.write('(')
+                        output.print_node(node.rarg)
+                        output.write(')')
+                    else:
+                        output.print_node(node.rarg)
         else:
             output.write('SELECT')
             if node.distinctClause:
@@ -776,7 +800,8 @@ def select_stmt(node, output):
                     output.print_list(node.distinctClause)
                     output.write(')')
             output.write(' ')
-            output.print_list(node.targetList)
+            if node.targetList:
+                output.print_list(node.targetList)
             if node.fromClause:
                 output.newline()
                 output.write('FROM ')